requirements.txt
# Core dependencies for BrainDance backend
ollama==0.4.7
openai>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
# qdrant-client>=1.8.0
mem0ai>=0.1.93
langchain-core>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
urllib3>=1.26.0,<2.0
httpx>=0.24.0
weaviate-client>=4.11.0

werkzeug==3.1.3
click==8.1.8
authlib==1.3.1
deprecation==2.1.0

# Additional useful packages
python-dotenv>=0.19.0
